Implement dark theme in options page.
Make auto theme react to system config changes.
