--- interface/options/dark.css
@@ -1,7 +1,0 @@
-@media (prefers-color-scheme: dark) {
-  /* TODO */
-  /*body {
-    background-color: #2c2e31;
-    color: #ffffff;
-  }*/
-}
