+++ interface/options/options.html
@@ -7,7 +7,6 @@
     <script src="options.js" type="module" defer></script>
     <link rel="stylesheet" href="style.css" />
     <link rel="stylesheet" href="switch.css" />
-    <link rel="stylesheet" href="dark.css" />
     <title>Cookie-Editor Options</title>
   </head>
   <body>
@@ -84,27 +83,35 @@
         </p>
         <p>
           View the
-          <a href="https://github.com/Moustachauve/cookie-editor">Source code</a
-          >.
-        </p>
-        <p>
-          Read the
-          <a href="https://cookie-editor.cgagnier.ca/privacy.html"
+          <a
+            href="https://github.com/Moustachauve/cookie-editor"
+            target="_blank"
+            >Source code</a
+          >. Read the
+          <a
+            href="https://cookie-editor.cgagnier.ca/privacy.html"
+            target="_blank"
             >Privacy Policy</a
-          >.
-        </p>
-        <p>
+          >. <br />
           Cookie-Editor is published under
           <a
             href="https://github.com/Moustachauve/cookie-editor/blob/master/LICENSE"
+            target="_blank"
             >the GPL-3.0 license</a
           >.
         </p>
         <p>
-          Are you enjoying Cookie-Editor? <br />
+          <em>Are you enjoying Cookie-Editor?</em><br />
           I would be really greatful if you considered supporting my work by
           sponsering me on Github.
-          <a href="https://github.com/sponsors/Moustachauve/">Learn more</a>.
+          <a href="https://github.com/sponsors/Moustachauve/" target="_blank"
+            >Learn more</a
+          >.
+        </p>
+        <p class="disclaimer">
+          This project is not an official Google project. It is not supported by
+          Google and Google specifically disclaims all warranties as to its
+          quality, merchantability, or fitness for a particular purpose.
         </p>
       </fieldset>
     </div>
