+++ interface/options/options.js
@@ -1,17 +1,20 @@
 import { BrowserDetector } from '../lib/browserDetector.js';
 import { GenericStorageHandler } from '../lib/genericStorageHandler.js';
 import { OptionsHandler } from '../lib/optionsHandler.js';
+import { ThemeHandler } from '../lib/themeHandler.js';
 
 document.addEventListener('DOMContentLoaded', async (event) => {
   const browserDetector = new BrowserDetector();
   const storageHandler = new GenericStorageHandler(browserDetector);
   const optionHandler = new OptionsHandler(browserDetector, storageHandler);
+  const themeHandler = new ThemeHandler(optionHandler);
   const advancedCookieInput = document.getElementById('advanced-cookie');
   const showDevtoolsInput = document.getElementById('devtool-show');
   const exportFormatInput = document.getElementById('export-format');
   const themeInput = document.getElementById('theme');
 
   await optionHandler.loadOptions();
+  themeHandler.updateTheme();
   optionHandler.on('optionsChanged', setFormValues);
   setFormValues();
   setInputEvents();
@@ -42,6 +45,7 @@ document.addEventListener('DOMContentLoaded', async (event) => {
     });
     themeInput.addEventListener('change', (_event) => {
       optionHandler.setTheme(themeInput.value);
+      themeHandler.updateTheme();
     });
   }
 });
