+++ interface/options/style.css
@@ -1,6 +1,11 @@
+@import url('../theme/light.css');
+@import url('../theme/dark.css');
+
 body {
   font-family: 'Segoe UI', Tahoma, sans-serif;
-  font-size: 100%; 
+  font-size: 100%;
+  background-color: var(--primary-surface-color);
+  color: var(--primary-text-color);
 }
 
 body * {
@@ -14,6 +19,14 @@ svg.icon {
   pointer-events: none;
 }
 
+a {
+  color: var(--primary-link-color);
+}
+
+a:hover {
+  color: var(--primary-link-hover-color);
+}
+
 .container {
   width: 100%;
   max-width: 600px;
@@ -23,7 +36,7 @@ svg.icon {
 .container > fieldset {
   margin: 15px;
   padding: 5px 15px;
-  border: 1px solid #bababa;
+  border: 1px solid var(--primary-border-color);
   border-radius: 4px;
 }
 .container > fieldset legend {
@@ -37,17 +50,28 @@ svg.icon {
   min-height: 2em;
 }
 
-.input-container input, .input-container select, .input-container .switch {
+.input-container input,
+.input-container select,
+.input-container .switch {
   float: right;
 }
 .input-container select {
   min-width: 120px;
+  background-color: var(--primary-surface-color);
+  color: var(--primary-text-color);
+  padding: 2px;
 }
 
 .hint {
   clear: both;
-  color: #555;
+  color: var(--secondary-text-color);
   font-size: 0.8em;
   padding: 1em;
   padding-top: 2px;
+}
+
+.disclaimer {
+  font-size: 0.8em;
+  color: var(--secondary-text-color);
+  padding-top: 5px;
 }
