+++ interface/popup-android/cookie-list.html
@@ -323,7 +323,7 @@
         You do not have the permission required to read the cookies for this
         page.
         <div>
-          Request permission for
+          Request permission for...
           <div class="button-container">
             <button
               id="request-permission"
