+++ interface/popup/cookie-list.html
@@ -5,7 +5,6 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="style.css" />
-    <link rel="stylesheet" href="dark.css" />
     <title>Cookie-Editor</title>
   </head>
 
@@ -335,7 +334,7 @@
         You do not have the permission required to read the cookies for this
         page.
         <div>
-          Request permission for
+          Request permission for...
           <div class="button-container">
             <button
               id="request-permission"
